kernel: catch infinite recursion comparing plists and precords
Fixes #1150
