+++ src/funcs.h
@@ -43,9 +43,20 @@ extern  void            ExecEnd (
 
 
 /* TL: extern Int RecursionDepth; */
+extern UInt RecursionTrapInterval;
+extern void RecursionDepthTrap( void );
+
+static inline void CheckRecursionBefore( void )
+{
+    TLS(RecursionDepth)++;
+    if ( RecursionTrapInterval &&
+         0 == (TLS(RecursionDepth) % RecursionTrapInterval) )
+      RecursionDepthTrap();
+}
+
+
 /****************************************************************************
 **
-
 *F * * * * * * * * * * * * * initialize package * * * * * * * * * * * * * * *
 */
