+++ tst/testbugfix/00351.tst
@@ -1,0 +1,10 @@
+# 2017-02-18 (MH): Comparing recursive data structures should not
+# crash. See issue #1150
+gap> [~] < [~];
+Error, recursion depth trap (5000)
+gap> [~] = [~];
+Error, recursion depth trap (5000)
+gap> rec(a:=~) = rec(a:=~);
+Error, recursion depth trap (5000)
+gap> rec(a:=~) < rec(a:=~);
+Error, recursion depth trap (5000)
